+++ docs/changelog.rst
@@ -16,12 +16,14 @@ Core Object
 Additions to the :doc:`core-object-methods`.
 
 * Added the :py:meth:`khoros.core.Khoros.User.get_all_users_count` method.
+* Added the :py:meth:`khoros.core.Khoros.BulkData.filter_by_action` static method.
 
 Primary Modules
 ---------------
 Additions to the :doc:`primary modules <primary-modules>`.
 
 * Added the :py:func:`khoros.objects.users.get_all_users_count` function.
+* Added the :py:func:`khoros.bulk_data.filter_by_action` function.
 
 Changed
 =======